--- src/main/java/faang/school/notificationservice/config/RedisConfiguration.java
@@ -1,44 +1,0 @@
-package faang.school.notificationservice.config;
-
-import faang.school.notificationservice.listener.LikeEventListener;
-import lombok.AllArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-
-@Configuration
-@AllArgsConstructor
-public class RedisConfiguration {
-    private final RedisProperties redisProperties;
-
-    @Bean
-    public LettuceConnectionFactory connectionFactory() {
-        RedisStandaloneConfiguration config =
-                new RedisStandaloneConfiguration(redisProperties.getHost(), redisProperties.getPort());
-        return new LettuceConnectionFactory(config);
-    }
-
-    @Bean
-    public ChannelTopic likeEventTopic() {
-        return new ChannelTopic(redisProperties.getChannel().getLike());
-    }
-
-    @Bean
-    public MessageListenerAdapter redisLikeEventListener(LikeEventListener likeEventListener) {
-        return new MessageListenerAdapter(likeEventListener);
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisMessageListenerContainer(MessageListenerAdapter redisLikeEventListener) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(connectionFactory());
-
-        container.addMessageListener(redisLikeEventListener, likeEventTopic());
-
-        return container;
-    }
-}
