+++ src/main/java/faang/school/notificationservice/dto/user/UserDto.java
@@ -2,6 +2,7 @@ package faang.school.notificationservice.dto.user;
 
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder(toBuilder = true)
