+++ src/main/java/faang/school/notificationservice/listener/like/LikeEventListener.java
@@ -1,9 +1,10 @@
-package faang.school.notificationservice.listener;
+package faang.school.notificationservice.listener.like;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
-import faang.school.notificationservice.dto.UserDto;
+import faang.school.notificationservice.dto.user.UserDto;
 import faang.school.notificationservice.dto.publishable.LikeEvent;
+import faang.school.notificationservice.listener.AbstractEventListenerByDima;
 import faang.school.notificationservice.messaging.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
 import org.springframework.stereotype.Component;
