+++ src/test/java/faang/school/notificationservice/listener/AbstractEventListenerByDimaTest.java
@@ -1,7 +1,7 @@
 package faang.school.notificationservice.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.notificationservice.dto.UserDto;
+import faang.school.notificationservice.dto.user.UserDto;
 import faang.school.notificationservice.messaging.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
 import org.junit.jupiter.api.BeforeEach;
@@ -30,7 +30,7 @@ public class AbstractEventListenerByDimaTest {
     private MessageBuilder<Object> messageBuilder;
     private Class<Object> eventType;
     private AbstractEventListenerByDima<Object> eventListener;
-    private UserDto userDto = new UserDto();
+    private UserDto userDto = UserDto.builder().build();
     private Object[] args = new Object[0];
     private Message message = new Message() {
         @Override
@@ -87,5 +87,4 @@ public class AbstractEventListenerByDimaTest {
 
         verify(notificationService).send(userDto, text);
     }
-
 }
